requirements.txt
streamlit
psycopg2-binary
pandas
fastapi
uvicorn
requests
openai
